abstract out collectParameters
